change variables

+++ src/main/java/com/softserveinc/softtour/entity/Food.java
@@ -10,7 +10,7 @@ public class Food {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private Integer id;
+    private long id;
 
     @Column(name = "name", nullable = false, length = 3)
     private String name;
@@ -25,11 +25,11 @@ public class Food {
         this.name = name;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
